refactor(models): promote models/attack.py to a models/attack/ package (#161)
One module per attack-surface formalism, so the agent boundary each artefact
crosses is legible from the import line - matching the models/asset/ package
shape and the AttackGraph / AttackTree / AttackForest framing in
academic-grounding.md.

- graph.py (AttackGraph - the OA's recon bundle, composing the OAM asset
  shapes), tree.py (AttackTree - the VR's per-goal hypothesis), forest.py
  (AttackForest plus its validation issue / report shapes and the
  finalisation error - all AttackForest's surface).
- Intra-package DAG: tree / graph leaves, forest -> tree. No model_rebuild.
- tests/models/test_attack.py mirrored into tests/models/attack/test_*.py.

Every consumer uses 'from models.attack import X' / 'from models import X',
both served by the package __init__ re-export - no import sites touched
outside the package. Pure relocation, no shape or behaviour change.
